Work better with more than one intermediate twist joint
Previously this script was fairly hard-coded to only allow for one
intermediate twist joint. Now it can *mostly* handle multiple joints
with a little more flexibility.

TODO: It still only creates one intermediate twist joint as it did
before. The createTwistJnts function just needs to be redone to allow
for more flexibility with the number of joints to create.

TODO: It also doesn't properly calculate the degree to which each joint
should rotate (like a constraint weight value). This issue is in the
driveJoints function on the multiplyDivide nodes.